+++ Source/Jeeves.Core.UnitTests/JeevesModuleTests.cs
@@ -12,7 +12,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void GetValue_Ok_UseHttp()
         {
-            var settings = new JeevesSettings(false, true);
+            var settings = CreateSettings(false, true);
             var store = CreateStoreWithUser();
             store.RetrieveValue("admin", "my_app", "my_key").Returns("{ \"Data\" : \"Foo\" }");
 
@@ -24,7 +24,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void GetValue_Ok_UseHttps()
         {
-            var settings = new JeevesSettings(true, true);
+            var settings = CreateSettings(true, true);
             var store = CreateStoreWithUser();
             store.RetrieveValue("admin", "my_app", "my_key").Returns("{ \"Data\" : \"Foo\" }");
 
@@ -40,7 +40,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void GetValue_Ok_NoAuthentication()
         {
-            var settings = new JeevesSettings(false, false);
+            var settings = CreateSettings(false, false);
             var store = Substitute.For<IDataStore>();
             store.RetrieveValue("admin", "my_app", "my_key").Returns("{ \"Data\" : \"Foo\" }");
 
@@ -55,7 +55,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void GetValue_SeeOther_NoHttps()
         {
-            var settings = new JeevesSettings(true, true);
+            var settings = CreateSettings(true, true);
             var store = CreateStoreWithUser();
 
             var response = PerformHttpGetRequest(settings, store, "/get/admin/my_app/my_key", "my_api");
@@ -66,7 +66,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void GetValue_Unauthorized_NoApikey()
         {
-            var settings = new JeevesSettings(false, true);
+            var settings = CreateSettings(false, true);
             var store = Substitute.For<IDataStore>();
 
             var response = PerformGetRequest(settings, store, "/get/admin/my_app/my_key", with => with.HttpRequest());
@@ -77,7 +77,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void GetValue_Unauthorized_WrongApikey()
         {
-            var settings = new JeevesSettings(false, true);
+            var settings = CreateSettings(false, true);
             var store = Substitute.For<IDataStore>();
 
             var response = PerformHttpGetRequest(settings, store, "/get/admin/my_app/my_key", "some_other_secret");
@@ -88,7 +88,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void GetValue_NoContent_NoValueInDataStore()
         {
-            var settings = new JeevesSettings(false, true);
+            var settings = CreateSettings(false, true);
             var store = CreateStoreWithUser();
 
             var response = PerformHttpGetRequest(settings, store, "/get/admin/my_app/my_key", "my_api");
@@ -99,7 +99,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void PostValue_Ok()
         {
-            var settings = new JeevesSettings(false, true);
+            var settings = CreateSettings(false, true);
             var store = CreateStoreWithUser();
 
             var response = PerformPostRequest(settings, store, "/post/admin/my_app/my_key", with =>
@@ -116,7 +116,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void PostValue_Ok_NoAuthentication()
         {
-            var settings = new JeevesSettings(false, false);
+            var settings = CreateSettings(false, false);
             var store = Substitute.For<IDataStore>();
 
             var response = PerformPostRequest(settings, store, "/post/admin/my_app/my_key", with =>
@@ -132,7 +132,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void PostValue_BadRequest_NoContent()
         {
-            var settings = new JeevesSettings(false, true);
+            var settings = CreateSettings(false, true);
             var store = CreateStoreWithUser();
 
             var response = PerformPostRequest(settings, store, "/post/admin/my_app/my_key", with =>
@@ -147,7 +147,7 @@ namespace Jeeves.Core.UnitTests
         [TestMethod]
         public void PostValue_BadRequest_EmptyContent()
         {
-            var settings = new JeevesSettings(false, true);
+            var settings = CreateSettings(false, true);
             var store = CreateStoreWithUser();
 
             var response = PerformPostRequest(settings, store, "/post/admin/my_app/my_key", with =>
@@ -165,6 +165,11 @@ namespace Jeeves.Core.UnitTests
             Assert.AreEqual("text/html", response.ContentType);
             Assert.AreEqual(HttpStatusCode.OK, response.StatusCode);
             Assert.AreEqual(expectedValue, response.Body.AsString());
+        }
+
+        private JeevesSettings CreateSettings(bool useHttps, bool useAuthentication)
+        {
+            return new JeevesSettings("some_uri", useHttps, useAuthentication);
         }
 
         private IDataStore CreateStoreWithUser()